UI and CRUD working

+++ src/controller/deleteFilm.java
@@ -1,8 +1,8 @@
 package controller;
 
 import java.io.IOException;
-import java.io.PrintWriter;
 
+import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -23,12 +23,22 @@ public class deleteFilm extends HttpServlet {
 
 
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+		response.setHeader("Cache-Control", "no-cache");
+	    response.setHeader("Pragma", "no-cache");
 		FilmDAO filmDAO = FilmDAO.getFilmDAO();
 		int userFilmID = Integer.parseInt(request.getParameter("ID"));
-		String title = filmDAO.getFilmByID(userFilmID).getFilmTitle();
-		filmDAO.deleteFilm(userFilmID);
-		PrintWriter out = response.getWriter();
-		out.println("deleted " + title);
+		String address = "";
+		if (filmDAO.getFilmByID(userFilmID) == null) {
+			request.setAttribute("filmId", userFilmID);
+			address = "delete-failed.jsp";
+		} else {
+			String title = filmDAO.getFilmByID(userFilmID).getFilmTitle();
+			request.setAttribute("filmTitle", title);
+			address = "delete-succeded.jsp";
+			filmDAO.deleteFilm(userFilmID);
+		}
+		RequestDispatcher dispatcher = request.getRequestDispatcher(address);
+	    dispatcher.include(request, response);
 	}
 
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
